New Index In Development
This file is still in early development. But when this is done, it will be the new Index page of this website. This will contain a card layout with a picture, basic information, social links, and more. It will then lead to other sections of the site, made of other cards that will hold portfolios of my code and design, as well as photography from Unsplash, and will also list out any gear that I use on an every day basis.